Rename io.buoyant.Linkerd to io.buoyant.linkerd.Main
Consistent with namerd. Makes sense in general.
